requirements.txt
openai
requests
beautifulsoup4
pyyaml
ddgs
defusedxml

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-asyncio>=0.21.0
freezegun>=1.2.0